Create Flasher.h
This Class allows the creation of Flashers, LED controllers that flash LEDs in some repeating pattern. The flasher is created with a call to the Constructor with a flashing pattern (ON times and OFF times in an array). The flasher can be controlled with calls to stop(), start(), pause(), resume(), and setFlash(), but once the flasher is created, the main program does not need to anything else except make one call to Flasher::loop() at the beginning of the main program loop(). Multiple flashers can be created. There is a test/sample program named TestFlasher.ino.