Create Keyboard.h
This Class enables the user to connect a matrix-style keyboard or keypad to their program simply. The keyboard is specified with matrix output pins and input pins, and a keymap string of characters. The keyboard object debounces keys, maps the keys to characters, repeats the keys if held, allows a limited amount of key-rollover and buffering, and enables the user to specify one shift-key that will give all the other keys a secondary (shifted) mapping. The main program needs to create the Keyboard object, call keyboard.loop(); at the top of loop(), and call getKey() to check for and retrieve keystrokes. There is a test and demonstration application in the file named TestKeyboard.ino.